+++ src/components/Feed/SchedulingFeed/ScheduleEditView.tsx
@@ -3,7 +3,8 @@ import { Button } from '@components/common/Button/Button';
 import Flex from '@components/common/Flex/Flex';
 import TimeColumn from '@components/Feed/SchedulingFeed/TimeColumn';
 import useScheduleSelection from '@hooks/feed/useScheduleSelection';
-import { isPastSlot, makeSlotId } from '@utils/feed/scheduling/schedulingUtils';
+import { isPastSlot } from '@utils/feed/scheduling/pastSlot';
+import { makeSlotId } from '@utils/feed/scheduling/slotId';
 import type { AvailabilityColumn } from '@type/feed';
 import * as S from './SchedulingFeed.styled';
